+++ internal/pkg/cmd/cli/cluster/delete.go
@@ -104,7 +104,7 @@ func (c *deleteCmd) run() error {
 		}
 	} else {
 		_, err = printer.Fprintf("[white][bold]Note: [reset]No destructive changes were made. To actually "+
-			"delete the cluster, rerun this command passing [cyan][bold]--%s[reset]\n", constants.YES_FLAG)
+			"delete the cluster, rerun this command passing [cyan][bold]--%s[reset]\n", constants.YesFlag)
 		if err != nil {
 			return errors.WithStack(err)
 		}